+++ src/libguac/socket.c
@@ -238,7 +238,7 @@ ssize_t guac_socket_write_string(guac_socket* socket, const char* str) {
 
 }
 
-ssize_t guac_socket_write_bytes(guac_socket* socket, const char* buf, unsigned int len) {
+ssize_t guac_socket_write_bytes(guac_socket* socket, const char* buf, int64_t len) {
 
     /* Write contents of string */
     if (guac_socket_write(socket, buf, len))
